change Vehicle.capacity type from double to int

+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/data/file/VehicleReader.java
@@ -33,7 +33,7 @@ import org.xml.sax.Attributes;
 public class VehicleReader extends MatsimXmlParser {
 	private static final String VEHICLE = "vehicle";
 
-	private static final double DEFAULT_CAPACITY = 1;
+	private static final int DEFAULT_CAPACITY = 1;
 	private static final double DEFAULT_T_0 = 0;
 	private static final double DEFAULT_T_1 = 24 * 60 * 60;
 
@@ -59,13 +59,13 @@ public class VehicleReader extends MatsimXmlParser {
 	private Vehicle createVehicle(Attributes atts) {
 		Id<Vehicle> id = Id.create(atts.getValue("id"), Vehicle.class);
 		Link startLink = links.get(Id.createLinkId(atts.getValue("start_link")));
-		double capacity = ReaderUtils.getDouble(atts, "capacity", DEFAULT_CAPACITY);
+		int capacity = ReaderUtils.getInt(atts, "capacity", DEFAULT_CAPACITY);
 		double t0 = ReaderUtils.getDouble(atts, "t_0", DEFAULT_T_0);
 		double t1 = ReaderUtils.getDouble(atts, "t_1", DEFAULT_T_1);
 		return createVehicle(id, startLink, capacity, t0, t1, atts);
 	}
 
-	protected Vehicle createVehicle(Id<Vehicle> id, Link startLink, double capacity, double t0, double t1,
+	protected Vehicle createVehicle(Id<Vehicle> id, Link startLink, int capacity, double t0, double t1,
 			Attributes atts) {
 		return new VehicleImpl(id, startLink, capacity, t0, t1);
 	}
